+++ pom.xml
@@ -41,7 +41,7 @@
 
     <properties>
         <antlr4.version>4.8-1</antlr4.version>
-        <cucumber-jvm.version>4.8.1</cucumber-jvm.version>
+        <cucumber-jvm.version>6.6.0</cucumber-jvm.version>
         <docker-java.version>3.2.5</docker-java.version>
         <java.version>11</java.version>
         <jersey.version>2.27</jersey.version>
@@ -200,13 +200,19 @@
 
         <dependency>
             <groupId>io.cucumber</groupId>
-            <artifactId>cucumber-junit</artifactId>
+            <artifactId>cucumber-java</artifactId>
             <version>${cucumber-jvm.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>io.cucumber</groupId>
             <artifactId>cucumber-spring</artifactId>
+            <version>${cucumber-jvm.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>io.cucumber</groupId>
+            <artifactId>cucumber-junit-platform-engine</artifactId>
             <version>${cucumber-jvm.version}</version>
             <scope>test</scope>
         </dependency>
@@ -232,6 +238,14 @@
     </dependencyManagement>
 
     <build>
+        <testResources>
+            <testResource>
+                <directory>src/test/resources/</directory>
+            </testResource>
+            <testResource>
+                <directory>src/test/features</directory>
+            </testResource>
+        </testResources>
         <pluginManagement>
             <plugins>
                 <plugin>
