+++ src/test/features/io/gaia_app/e2e/basic_navigation.feature
@@ -1,3 +1,4 @@
+@e2e
 Feature: Basic Navigation
 
     Background:
