+++ src/test/java/io/gaia_app/e2e/CucumberIT.java
@@ -1,15 +1,7 @@
 package io.gaia_app.e2e;
 
-import io.cucumber.junit.Cucumber;
-import io.cucumber.junit.CucumberOptions;
-import org.junit.jupiter.api.Tag;
-import org.junit.runner.RunWith;
+import io.cucumber.junit.platform.engine.Cucumber;
 
-@RunWith(Cucumber.class)
-@CucumberOptions(plugin = "pretty", features = "src/test/features")
-@Tag("e2e")
+@Cucumber
 public class CucumberIT {
-
-    // ces TU sont exécutés avec JUnit 4, et ne sont donc pas filtrés par maven avec le tag
-
 }
