+++ src/test/resources/junit-platform.properties
@@ -1,1 +1,3 @@
 junit.jupiter.execution.parallel.enabled=true
+
+cucumber.publish.quiet=true
